Extract Mac/Windows NetworkParams common bases
Mac and Windows were the two platforms whose NetworkParams still lacked a
common base (unlike Linux/AIX/BSD/Solaris), so their JNA and FFM twins each
duplicated the command-line default-gateway lookups. Add the missing bases and
hoist that shared logic.

MacNetworkParams holds getIpv4/getIpv6DefaultGateway (route / netstat parsing);
the Mac twins keep only their native getDomainName/getHostName (getaddrinfo /
gethostname). WindowsNetworkParams holds getIpv4/getIpv6DefaultGateway (route
print parsing); the Windows twins keep native getDomainName/getDnsServers/
getHostName. The twins' super.getHostName() fallbacks still resolve to
AbstractNetworkParams, which the new bases do not override.

No user-facing change; NetworkParams JNA==FFM comparison passes.
