+++ oshi-common/src/main/java/oshi/software/common/os/mac/MacNetworkParams.java
@@ -1,0 +1,50 @@
+/*
+ * Copyright 2017-2026 The OSHI Project Contributors
+ * SPDX-License-Identifier: MIT
+ */
+package oshi.software.common.os.mac;
+
+import java.util.List;
+
+import oshi.annotation.concurrent.ThreadSafe;
+import oshi.software.common.AbstractNetworkParams;
+import oshi.util.ExecutingCommand;
+import oshi.util.ParseUtil;
+
+/**
+ * Common Mac NetworkParams logic shared between the JNA and FFM implementations. The default-gateway lookups are
+ * command-line based and live here; hostname/domain resolution is native and provided by the subclasses.
+ */
+@ThreadSafe
+public abstract class MacNetworkParams extends AbstractNetworkParams {
+
+    private static final String IPV6_ROUTE_HEADER = "Internet6:";
+
+    private static final String DEFAULT_GATEWAY = "default";
+
+    /** Default constructor. */
+    protected MacNetworkParams() {
+    }
+
+    @Override
+    public String getIpv4DefaultGateway() {
+        return searchGateway(ExecutingCommand.runNative("route -n get default"));
+    }
+
+    @Override
+    public String getIpv6DefaultGateway() {
+        List<String> lines = ExecutingCommand.runNative("netstat -nr");
+        boolean v6Table = false;
+        for (String line : lines) {
+            if (v6Table && line.startsWith(DEFAULT_GATEWAY)) {
+                String[] fields = ParseUtil.whitespaces.split(line);
+                if (fields.length > 2 && fields[2].contains("G")) {
+                    return fields[1].split("%")[0];
+                }
+            } else if (line.startsWith(IPV6_ROUTE_HEADER)) {
+                v6Table = true;
+            }
+        }
+        return "";
+    }
+}
