+++ oshi-common/src/main/java/oshi/software/common/os/windows/WindowsNetworkParams.java
@@ -1,0 +1,57 @@
+/*
+ * Copyright 2017-2026 The OSHI Project Contributors
+ * SPDX-License-Identifier: MIT
+ */
+package oshi.software.common.os.windows;
+
+import java.util.List;
+
+import oshi.annotation.concurrent.ThreadSafe;
+import oshi.software.common.AbstractNetworkParams;
+import oshi.util.ExecutingCommand;
+import oshi.util.ParseUtil;
+
+/**
+ * Common Windows NetworkParams logic shared between the JNA and FFM implementations. The default-gateway lookups parse
+ * {@code route print} output and live here; domain/host name and DNS-server resolution are native and provided by the
+ * subclasses.
+ */
+@ThreadSafe
+public abstract class WindowsNetworkParams extends AbstractNetworkParams {
+
+    /** Default constructor. */
+    protected WindowsNetworkParams() {
+    }
+
+    @Override
+    public String getIpv4DefaultGateway() {
+        return parseIpv4Route();
+    }
+
+    @Override
+    public String getIpv6DefaultGateway() {
+        return parseIpv6Route();
+    }
+
+    private static String parseIpv4Route() {
+        List<String> lines = ExecutingCommand.runNative("route print -4 0.0.0.0");
+        for (String line : lines) {
+            String[] fields = ParseUtil.whitespaces.split(line.trim());
+            if (fields.length > 2 && "0.0.0.0".equals(fields[0])) {
+                return fields[2];
+            }
+        }
+        return "";
+    }
+
+    private static String parseIpv6Route() {
+        List<String> lines = ExecutingCommand.runNative("route print -6 ::/0");
+        for (String line : lines) {
+            String[] fields = ParseUtil.whitespaces.split(line.trim());
+            if (fields.length > 3 && "::/0".equals(fields[2])) {
+                return fields[3];
+            }
+        }
+        return "";
+    }
+}
