+++ oshi-core-ffm/src/main/java/oshi/software/os/windows/WindowsNetworkParamsFFM.java
@@ -4,15 +4,11 @@
  */
 package oshi.software.os.windows;
 
-import java.util.List;
-
 import oshi.ffm.util.platform.windows.IPHlpAPIUtilFFM;
 import oshi.ffm.util.platform.windows.Kernel32UtilFFM;
-import oshi.software.common.AbstractNetworkParams;
-import oshi.util.ExecutingCommand;
-import oshi.util.ParseUtil;
+import oshi.software.common.os.windows.WindowsNetworkParams;
 
-public final class WindowsNetworkParamsFFM extends AbstractNetworkParams {
+public final class WindowsNetworkParamsFFM extends WindowsNetworkParams {
 
     @Override
     public String[] getDnsServers() {
@@ -28,37 +24,5 @@ public final class WindowsNetworkParamsFFM extends AbstractNetworkParams {
     public String getHostName() {
         String name = Kernel32UtilFFM.getComputerName();
         return name.isEmpty() ? super.getHostName() : name;
-    }
-
-    @Override
-    public String getIpv4DefaultGateway() {
-        return parseIpv4Route();
-    }
-
-    @Override
-    public String getIpv6DefaultGateway() {
-        return parseIpv6Route();
-    }
-
-    private static String parseIpv4Route() {
-        List<String> lines = ExecutingCommand.runNative("route print -4 0.0.0.0");
-        for (String line : lines) {
-            String[] fields = ParseUtil.whitespaces.split(line.trim());
-            if (fields.length > 2 && "0.0.0.0".equals(fields[0])) {
-                return fields[2];
-            }
-        }
-        return "";
-    }
-
-    private static String parseIpv6Route() {
-        List<String> lines = ExecutingCommand.runNative("route print -6 ::/0");
-        for (String line : lines) {
-            String[] fields = ParseUtil.whitespaces.split(line.trim());
-            if (fields.length > 3 && "::/0".equals(fields[2])) {
-                return fields[3];
-            }
-        }
-        return "";
     }
 }
