+++ oshi-core/src/main/java/oshi/software/os/mac/MacNetworkParamsJNA.java
@@ -8,7 +8,6 @@ import static com.sun.jna.platform.unix.LibCAPI.HOST_NAME_MAX;
 
 import java.net.InetAddress;
 import java.net.UnknownHostException;
-import java.util.List;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -20,23 +19,17 @@ import oshi.jna.ByRef.CloseablePointerByReference;
 import oshi.jna.platform.mac.SystemB;
 import oshi.jna.platform.unix.CLibrary;
 import oshi.jna.platform.unix.CLibrary.Addrinfo;
-import oshi.software.common.AbstractNetworkParams;
-import oshi.util.ExecutingCommand;
-import oshi.util.ParseUtil;
+import oshi.software.common.os.mac.MacNetworkParams;
 
 /**
  * MacNetworkParamsJNA class.
  */
 @ThreadSafe
-final class MacNetworkParamsJNA extends AbstractNetworkParams {
+final class MacNetworkParamsJNA extends MacNetworkParams {
 
     private static final Logger LOG = LoggerFactory.getLogger(MacNetworkParamsJNA.class);
 
     private static final SystemB SYS = SystemB.INSTANCE;
-
-    private static final String IPV6_ROUTE_HEADER = "Internet6:";
-
-    private static final String DEFAULT_GATEWAY = "default";
 
     @Override
     public String getDomainName() {
@@ -75,27 +68,5 @@ final class MacNetworkParamsJNA extends AbstractNetworkParams {
             return super.getHostName();
         }
         return Native.toString(hostnameBuffer);
-    }
-
-    @Override
-    public String getIpv4DefaultGateway() {
-        return searchGateway(ExecutingCommand.runNative("route -n get default"));
-    }
-
-    @Override
-    public String getIpv6DefaultGateway() {
-        List<String> lines = ExecutingCommand.runNative("netstat -nr");
-        boolean v6Table = false;
-        for (String line : lines) {
-            if (v6Table && line.startsWith(DEFAULT_GATEWAY)) {
-                String[] fields = ParseUtil.whitespaces.split(line);
-                if (fields.length > 2 && fields[2].contains("G")) {
-                    return fields[1].split("%")[0];
-                }
-            } else if (line.startsWith(IPV6_ROUTE_HEADER)) {
-                v6Table = true;
-            }
-        }
-        return "";
     }
 }
