+++ oshi-core/src/main/java/oshi/software/os/windows/WindowsNetworkParamsJNA.java
@@ -23,15 +23,13 @@ import com.sun.jna.platform.win32.WinError;
 
 import oshi.annotation.concurrent.ThreadSafe;
 import oshi.jna.ByRef.CloseableIntByReference;
-import oshi.software.common.AbstractNetworkParams;
-import oshi.util.ExecutingCommand;
-import oshi.util.ParseUtil;
+import oshi.software.common.os.windows.WindowsNetworkParams;
 
 /**
  * WindowsNetworkParamsJNA class.
  */
 @ThreadSafe
-final class WindowsNetworkParamsJNA extends AbstractNetworkParams {
+final class WindowsNetworkParamsJNA extends WindowsNetworkParams {
 
     private static final Logger LOG = LoggerFactory.getLogger(WindowsNetworkParamsJNA.class);
 
@@ -91,37 +89,5 @@ final class WindowsNetworkParamsJNA extends AbstractNetworkParams {
         } catch (Win32Exception e) {
             return super.getHostName();
         }
-    }
-
-    @Override
-    public String getIpv4DefaultGateway() {
-        return parseIpv4Route();
-    }
-
-    @Override
-    public String getIpv6DefaultGateway() {
-        return parseIpv6Route();
-    }
-
-    private static String parseIpv4Route() {
-        List<String> lines = ExecutingCommand.runNative("route print -4 0.0.0.0");
-        for (String line : lines) {
-            String[] fields = ParseUtil.whitespaces.split(line.trim());
-            if (fields.length > 2 && "0.0.0.0".equals(fields[0])) {
-                return fields[2];
-            }
-        }
-        return "";
-    }
-
-    private static String parseIpv6Route() {
-        List<String> lines = ExecutingCommand.runNative("route print -6 ::/0");
-        for (String line : lines) {
-            String[] fields = ParseUtil.whitespaces.split(line.trim());
-            if (fields.length > 3 && "::/0".equals(fields[2])) {
-                return fields[3];
-            }
-        }
-        return "";
     }
 }
